Move CustomSizeSecret to common place
